refactor: Reorganise files to reduce screenshot test temp file name length
